Changed member type to integer.

+++ ClosedRage.R
@@ -1,6 +1,6 @@
-setClass("ClosedRange", slots = list(start = "numeric", end = "numeric"))
+setClass("ClosedRange", slots = list(start = "integer", end = "integer"))
 
 testthat::test_that("Can create ClosedRange object", {
-  closed_range <- new("ClosedRange", start = 1, end = 2)
+  closed_range <- new("ClosedRange", start = 1L, end = 2L)
   testthat::expect_true(!is.null(closed_range))
 })
